% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capesdata}
\alias{capesdata}
\title{CAPES ranking of scientific journals}
\format{
data frame format as read by function \code{\link{read.csv}}.
}
\source{
\url{https://sucupira.capes.gov.br/sucupira/public/consultas/coleta/veiculoPublicacaoQualis/listaConsultaGeralPeriodicos.jsf}
}
\usage{
data(capesdata)
}
\description{
Table with classification of the scientific journals according
to \href{https://www.gov.br/capes/pt-br}{CAPES} strata
}
