% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_year.R
\name{filter_year}
\alias{filter_year}
\title{Filter the year of lattes information}
\usage{
filter_year(lattes_data, year)
}
\arguments{
\item{lattes_data}{output result from get_lattes_folder function.}

\item{year}{the year to filter the data for.}
}
\value{
list of tibbles filtered by the specified year
}
\description{
It filters the data obtained from get_lattes_folder to a user speficified year
}
\details{
The function will not filter education, basic information,
idioms, and fields of research(areas).
}
\examples{
folder_path <- system.file("lattes", package = "NUPEX")
lattes_folder_data <- get_lattes_folder(folder_path)
lattes_filtered <- filter_year(lattes_folder_data, year = 2020)
summary_lattes <- summary_all(lattes_filtered)
}
