% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_altmetric.R
\name{get_altmetric}
\alias{get_altmetric}
\title{Get papers altmetric data}
\usage{
get_altmetric(papers)
}
\arguments{
\item{papers}{table (tibble format) depicting papers published or accepted as obtained
from get_lattes or get_lattes_folder}
}
\value{
A table in tibble format.
}
\description{
It obtains altmetric paper data from a tibble obtained using
get_lattes or get_lattes_folder.
}
\examples{
\dontrun{
path_lattes <- paste0(system.file("lattes", package = "NUPEX"),
                     "/lattes2.xml")
lattes_data <- get_lattes(path_lattes)
alt <- get_altmetric(lattes_data$papers)
}
}
