% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lattes.R
\name{get_lattes}
\alias{get_lattes}
\title{Get all individual lattes information}
\usage{
get_lattes(file)
}
\arguments{
\item{file}{the name of XML file obtained from the lattes platform}
}
\value{
A list of tibbles
}
\description{
It summarizes individual XML lattes data into a list of tables
(in tibble format) separated by the type of information.
}
\examples{
path_lattes <- paste0(system.file("lattes", package = "NUPEX"),
 "/lattes1.xml")
lattes_data <- get_lattes(path_lattes)

}
