% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listcompID.R
\name{listcompID}
\alias{listcompID}
\title{Produce list of complete personal information}
\usage{
listcompID(lattesdata, collabs = NULL)
}
\arguments{
\item{lattesdata}{table (tibble format) derived by \code{\link{get_lattes}} or
\code{\link{get_lattes_folder}}, that will be used to extract a list of
complete personal information}

\item{collabs}{vector containing the complete names of collaborators for the
Graduate Program; the provided names have to be exactly as shown in their
Lattes profiles; the non-listed collaborators will automatically have their
"VINCULO" defined as "Permanente"}
}
\value{
Table in data frame format
}
\description{
It uses a tibble obtained with \code{\link{get_lattes}} or
\code{\link{get_lattes_folder}} to produce a data frame listing the complete
personal information (e.g. ORCID ID, Lattes ID, Nationality, institutional
address, PhD thesis), as publicily available in the
\href{http://lattes.cnpq.br/}{Plataforma Lattes}
}
\examples{
\dontrun{
path_lattes <- paste0(system.file("lattes", package = "NUPEX"),
                      "/lattes2.xml")
lattes_data <- get_lattes_folder(path_lattes)
lcompID <- listcompID(lattes_data)

path_lattes_folder <- system.file("lattes", package = "NUPEX")
lattes_folder_data <- get_lattes_folder(path_lattes_folder)
collaborators <- c("Suzana Telles da Cunha Lima",
                   "Cláudia Dias de Santana")
lcompIDs <- listcompID(lattes_folder_data,
                       collabs = collaborators)
}

}
\author{
Domingos Cardoso
}
