% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listprizes.R
\name{listprizes}
\alias{listprizes}
\title{Produce list of prizes}
\usage{
listprizes(lattesdata, quadre = NULL)
}
\arguments{
\item{lattesdata}{table (tibble format) derived by \code{\link{get_lattes}} or
\code{\link{get_lattes_folder}}, that will be used to extract a list of
prizes or award titles}

\item{quadre}{a vector defining a time frame of years to extract the list of
prizes or award titles; for example, a quadrennium like c(2017, 2020), which
is useful for CAPES Sucupira report}
}
\value{
Table in data frame format
}
\description{
It produces a data frame listing prizes or award titles, from
a tibble obtained using \code{\link{get_lattes}} or \code{\link{get_lattes_folder}}
}
\examples{
\dontrun{
path_lattes <- paste0(system.file("lattes", package = "NUPEX"),
                      "/lattes2.xml")
lattes_data <- get_lattes(path_lattes)
lprizes <- listprizes(lattes_data
                      quadre = c(2017, 2020))

path_lattes_folder <- system.file("lattes", package = "NUPEX")
lattes_folder_data <- get_lattes_folder(path_lattes_folder)
lprizes <- listprizes(lattes_folder_data,
                      quadre = c(2017, 2020))
}

}
\author{
Domingos Cardoso
}
