% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listrev.R
\name{listrev}
\alias{listrev}
\title{Produce list of contributions in editorial board or as referees for scientific journals}
\usage{
listrev(lattesdata, paperdata = NULL, quadre = NULL)
}
\arguments{
\item{lattesdata}{table (tibble format) derived by \code{\link{get_lattes}} or
\code{\link{get_lattes_folder}}, that will be used to extract a list of
contributing editorial board or referees}

\item{paperdata}{table (data frame format) derived by \code{\link{listpapers}}
function, which if provided may add the journal classification according to
CAPES strata, depending on how the list of paper production was generated}

\item{quadre}{vector defining a time frame of years to extract the list of
contributing editorial board or referees}
}
\value{
Table in data frame format
}
\description{
It produces a data frame listing the scientific journals that have
contributing editorial board or referees, from a tibble obtained using
\code{\link{get_lattes}} or \code{\link{get_lattes_folder}}
}
\examples{
\dontrun{
path_lattes <- paste0(system.file("lattes", package = "NUPEX"),
                      "/lattes2.xml")
lattes_data <- get_lattes(path_lattes)
data(capes_qualis)
lpapers <- listpapers(lattes_data,
                      capesdata = capes_qualis,
                      quadre = c(2017, 2020))
lpats <- listrev(lattes_data,
                 paperdata = lpapers
                 quadre = c(2017, 2020))

path_lattes_folder <- system.file("lattes", package = "NUPEX")
lattes_folder_data <- get_lattes_folder(path_lattes_folder)
data(capes_qualis)
lpapers <- listpapers(lattes_folder_data,
                      capesdata = capes_qualis,
                      quadre = c(2017, 2020))
lrev <- listrev(lattes_folder_data,
                 paperdata = lpapers,
                 quadre = c(2017, 2020))
}

}
\author{
Domingos Cardoso
}
