% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listtech.R
\name{listtech}
\alias{listtech}
\title{Produce list of technical works}
\usage{
listtech(nupexfile, quadre = NULL)
}
\arguments{
\item{nupexfile}{table (tibble format) derived by \code{\link{get_lattes}} or
\code{\link{get_lattes_folder}}, that will be used to extract a list of
technical works}

\item{quadre}{a vector defining a time frame of years to extract the list of
technical works; for example, a quadrennium like c(2017, 2020), which is useful
for CAPES Sucupira report}
}
\value{
Table in dataframe format
}
\description{
It produces a data frame listing the technical works, from a tibble
obtained using \code{\link{get_lattes}} or \code{\link{get_lattes_folder}}
}
\examples{
\dontrun{
path_lattes <- paste0(system.file("lattes", package = "NUPEX"),
                      "/lattes2.xml")
lattes_data <- get_lattes_folder(path_lattes)
ltech <- listtech(lattes_data,
                  quadre = c(2017, 2020))

path_lattes_folder <- system.file("lattes", package = "NUPEX")
lattes_folder_data <- get_lattes_folder(path_lattes_folder)
ltech <- listtech(lattes_folder_data,
                  quadre = c(2017, 2020))
}

}
\author{
Domingos Cardoso
}
