% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppgsuperv.R
\name{ppgsuperv}
\alias{ppgsuperv}
\title{Produce list of graduate programs where each one contributes with ongoing supervising}
\usage{
ppgsuperv(nupexfile, collabs = NULL)
}
\arguments{
\item{nupexfile}{table (tibble format) derived by \code{\link{get_lattes}} or
\code{\link{get_lattes_folder}}, that will be used to extract a list of
ongoing supervisions of MSc or PhD students at different graduate programs}

\item{collabs}{vector containing the complete names of collaborators for the
Graduate Program; the provided names have to be exactly as shown in their
Lattes profiles; the non-listed collaborators will automatically have their
"VINCULO" defined as "Permanente"}
}
\value{
Table in data frame format
}
\description{
It uses a tibble obtained with \code{\link{get_lattes}} or
\code{\link{get_lattes_folder}} to produce a data frame listing all graduate
programs where each one contributes with ongoing supervising of MSc or PhD students
}
\examples{
\dontrun{
path_lattes <- paste0(system.file("lattes", package = "NUPEX"),
                      "/lattes2.xml")
lattes_data <- get_lattes_folder(path_lattes)
ppgs <- ppgsuperv(lattes_data)

path_lattes_folder <- system.file("lattes", package = "NUPEX")
lattes_folder_data <- get_lattes_folder(path_lattes_folder)
collaborators <- c("Suzana Telles da Cunha Lima",
                   "Cláudia Dias de Santana")
ppgs <- ppgsuperv(lattes_folder_data,
                  collabs = collaborators)
}

}
\author{
Domingos Cardoso
}
