% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalpapers.R
\name{totalpapers}
\alias{totalpapers}
\title{Produce total paper production}
\usage{
totalpapers(paperdata, lattesdata, collabs = NULL)
}
\arguments{
\item{paperdata}{table (data frame format) derived by \code{\link{listpapers}}
function, which has to be used so as to get the total number of published papers}

\item{lattesdata}{table (tibble format) derived by \code{\link{get_lattes}} or
\code{\link{get_lattes_folder}}}

\item{collabs}{vector containing the complete names of collaborators for the
Graduate Program; the provided names have to be exactly as shown in their
Lattes profiles; the non-listed collaborators will automatically have their
"VINCULO" defined as "Permanente"}
}
\value{
Table in data frame format
}
\description{
It uses the list of publication (data frame format) derived by
\code{\link{listpapers}} so as to produce a summary of total paper production.
If the capesdata file was provided when retrieving the complete list of papers,
then this function will also return the total paper production in each stratum
of CAPES journal classification, as well as the total percentil
}
\examples{
\dontrun{
path_lattes <- paste0(system.file("lattes", package = "NUPEX"),
                      "/lattes2.xml")
lattes_data <- get_lattes(path_lattes)
lpapers <- listpapers(lattes_data,
                      capesdata = capes_qualis,
                      quadre = c(2017, 2020))
tpapers <- totalpapers(lpapers,
                       lattes_data)

path_lattes_folder <- system.file("lattes", package = "NUPEX")
lattes_folder_data <- get_lattes_folder(path_lattes_folder)
data(capes_qualis)
lpapers <- listpapers(lattes_folder_data,
                      capesdata = capes_qualis,
                      quadre = c(2017, 2020))
collaborators <- c("Suzana Telles da Cunha Lima",
                   "Cláudia Dias de Santana")
tpapers <- totalpapers(lpapers,
                       lattes_folder_data,
                       collabs = collaborators)
}

}
\author{
Domingos Cardoso
}
