% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listpapers.R
\name{listpapers}
\alias{listpapers}
\title{Get list of papers}
\usage{
listpapers(nupexfile, capesfile = NULL, quadre = NULL)
}
\arguments{
\item{nupexfile}{table (tibble format) which will be used to extract
published and accepted papers as obtained by \code{\link{get_lattes}} or
\code{\link{get_lattes_folder}}}

\item{capesfile}{a dataframe of journal classification according to CAPES strata}

\item{quadre}{a vector defing an interval of years to extract the list of papers}
}
\value{
A table in dataframe format.
}
\description{
It obtains a dataframe with published and accepted papers from a tibble obtained using
get_lattes or get_lattes_folder. By proviving a capes file, the function adds
the classification of each journal by CAPES strata
}
\examples{
\dontrun{
path_lattes <- paste0(system.file("lattes", package = "NUPEX"),
                     "/lattes2.xml")
lattes_data <- get_lattes(path_lattes)
data(capes_qualis)
lpapers <- listpapers(lattes_data
                      capesfile = capes_qualis,
                      quadre = c(2017, 2020))
}

}
\author{
Domingos Cardoso
}
